Move Atom API bridge components to lib/atom
