+++ lib/atom/atom-text-editor.js
