+++ lib/atom/commands.js
