+++ lib/atom/decoration.js
