+++ lib/atom/dock-item.js
