+++ lib/atom/octicon.js
