+++ lib/atom/panel.js
