+++ lib/atom/simple-tooltip.js
