+++ lib/atom/status-bar.js
