+++ lib/atom/tooltip.js
