+++ lib/containers/issueish-pane-item-container.js
@@ -7,7 +7,7 @@ import {autobind} from 'core-decorators';
 import IssueTimelineContainer from './issue-timeline-container';
 import PrTimelineContainer from './pr-timeline-container';
 import PrStatusesContainer from './pr-statuses-container';
-import Octicon from '../views/octicon';
+import Octicon from '../atom/octicon';
 import IssueishBadge from '../views/issueish-badge';
 import GithubDotcomMarkdown from '../views/github-dotcom-markdown';
 import PeriodicRefresher from '../periodic-refresher';
