+++ lib/containers/issueish-tooltip-container.js
@@ -3,7 +3,7 @@ import {createFragmentContainer, graphql} from 'react-relay';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
 
-import Octicon from '../views/octicon';
+import Octicon from '../atom/octicon';
 
 const typeAndStateToIcon = {
   Issue: {
