+++ lib/containers/pr-info-container.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import cx from 'classnames';
 import {autobind} from 'core-decorators';
 
-import Octicon from '../views/octicon';
+import Octicon from '../atom/octicon';
 import IssueishBadge from '../views/issueish-badge';
 import PeriodicRefresher from '../periodic-refresher';
 import tinycolor from 'tinycolor2';
