+++ lib/containers/pr-selection-by-branch-container.js
@@ -5,7 +5,7 @@ import {autobind} from 'core-decorators';
 
 import PrInfoContainer from './pr-info-container';
 import PrUrlInputBox from '../views/pr-url-input-box';
-import Octicon from '../views/octicon';
+import Octicon from '../atom/octicon';
 import {RemotePropType, BranchSetPropType} from '../prop-types';
 
 export class PrSelectionByBranch extends React.Component {
