+++ lib/containers/pr-status-context-container.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {createFragmentContainer, graphql} from 'react-relay';
 import PropTypes from 'prop-types';
 
-import Octicon from '../views/octicon';
+import Octicon from '../atom/octicon';
 import {stateToIconAndStyle} from './pr-statuses-container';
 
 export class PrStatusContext extends React.Component {
