+++ lib/containers/pr-statuses-container.js
@@ -5,7 +5,7 @@ import {autobind} from 'core-decorators';
 
 import {toSentence} from '../helpers';
 import PrStatusContextContainer from './pr-status-context-container';
-import Octicon from '../views/octicon';
+import Octicon from '../atom/octicon';
 import DonutChart from '../views/donut-chart';
 import PeriodicRefresher from '../periodic-refresher';
 import {RelayConnectionPropType} from '../prop-types';
