+++ lib/containers/timeline-items/cross-referenced-event-container.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {graphql, createFragmentContainer} from 'react-relay';
 import PropTypes from 'prop-types';
 
-import Octicon from '../../views/octicon';
+import Octicon from '../../atom/octicon';
 import IssueishBadge from '../../views/issueish-badge';
 import IssueishLink from '../../views/issueish-link';
