+++ lib/containers/timeline-items/cross-referenced-events-container.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {graphql, createFragmentContainer} from 'react-relay';
 import PropTypes from 'prop-types';
 
-import Octicon from '../../views/octicon';
+import Octicon from '../../atom/octicon';
 import Timeago from '../../views/timeago';
 import CrossReferencedEventContainer from './cross-referenced-event-container';
