+++ lib/containers/user-mention-tooltip-container.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {graphql, createFragmentContainer} from 'react-relay';
 import PropTypes from 'prop-types';
 
-import Octicon from '../views/octicon';
+import Octicon from '../atom/octicon';
 
 class UserMentionTooltip extends React.Component {
   static propTypes = {
