+++ lib/controllers/conflict-controller.js
@@ -5,8 +5,8 @@ const {Menu, MenuItem} = remote;
 
 import {autobind} from 'core-decorators';
 import {OURS, BASE, THEIRS} from '../models/conflicts/source';
-import Decoration from '../views/decoration';
-import Octicon from '../views/octicon';
+import Decoration from '../atom/decoration';
+import Octicon from '../atom/octicon';
 
 export default class ConflictController extends React.Component {
   static propTypes = {
