+++ lib/controllers/editor-conflict-controller.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import {autobind} from 'core-decorators';
 import compareSets from 'compare-sets';
 
-import Commands, {Command} from '../views/commands';
+import Commands, {Command} from '../atom/commands';
 import Conflict from '../models/conflicts/conflict';
 import ConflictController from './conflict-controller';
 import {OURS, THEIRS, BASE} from '../models/conflicts/source';
