+++ lib/controllers/root-controller.js
@@ -6,15 +6,15 @@ import PropTypes from 'prop-types';
 import {autobind} from 'core-decorators';
 import {CompositeDisposable} from 'event-kit';
 
-import StatusBar from '../views/status-bar';
-import Panel from '../views/panel';
+import StatusBar from '../atom/status-bar';
+import Panel from '../atom/panel';
 import PaneItem from '../atom/pane-item';
-import DockItem from '../views/dock-item';
+import DockItem from '../atom/dock-item';
 import CloneDialog from '../views/clone-dialog';
 import OpenIssueishDialog from '../views/open-issueish-dialog';
 import InitDialog from '../views/init-dialog';
 import CredentialDialog from '../views/credential-dialog';
-import Commands, {Command} from '../views/commands';
+import Commands, {Command} from '../atom/commands';
 import GitTimingsView from '../views/git-timings-view';
 import GithubTabController from './github-tab-controller';
 import FilePatchController from './file-patch-controller';
