+++ lib/controllers/status-bar-tile-controller.js
@@ -5,8 +5,8 @@ import BranchView from '../views/branch-view';
 import BranchMenuView from '../views/branch-menu-view';
 import PushPullView from '../views/push-pull-view';
 import ChangedFilesCountView from '../views/changed-files-count-view';
-import Tooltip from '../views/tooltip';
-import Commands, {Command} from '../views/commands';
+import Tooltip from '../atom/tooltip';
+import Commands, {Command} from '../atom/commands';
 import ObserveModel from '../views/observe-model';
 import RefHolder from '../models/ref-holder';
 import yubikiri from 'yubikiri';
