+++ lib/views/branch-menu-view.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import {autobind} from 'core-decorators';
 import cx from 'classnames';
 
-import Commands, {Command} from './commands';
+import Commands, {Command} from '../atom/commands';
 import {BranchPropType, BranchSetPropType} from '../prop-types';
 import {GitError} from '../git-shell-out-strategy';
