+++ lib/views/changed-files-count-view.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Octicon from './octicon';
+import Octicon from '../atom/octicon';
 
 export default class ChangedFilesCountView extends React.Component {
   static propTypes = {
