+++ lib/views/clone-dialog.js
@@ -5,7 +5,7 @@ import {CompositeDisposable} from 'event-kit';
 import url from 'url';
 import path from 'path';
 
-import Commands, {Command} from './commands';
+import Commands, {Command} from '../atom/commands';
 
 export default class CloneDialog extends React.Component {
   static propTypes = {
