+++ lib/views/co-author-form.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {autobind} from 'core-decorators';
 
-import Commands, {Command} from './commands';
+import Commands, {Command} from '../atom/commands';
 
 export default class CoAuthorForm extends React.Component {
   static propTypes = {
