+++ lib/views/commit-view.js
@@ -6,8 +6,8 @@ import cx from 'classnames';
 import Select from 'react-select';
 import CoAuthorForm from './co-author-form';
 
-import Tooltip from './tooltip';
-import AtomTextEditor from './atom-text-editor';
+import Tooltip from '../atom/tooltip';
+import AtomTextEditor from '../atom/atom-text-editor';
 import RefHolder from '../models/ref-holder';
 import {LINE_ENDING_REGEX} from '../helpers';
 import {AuthorPropType} from '../prop-types';
