+++ lib/views/file-patch-view.js
@@ -6,8 +6,8 @@ import cx from 'classnames';
 import {autobind} from 'core-decorators';
 
 import HunkView from './hunk-view';
-import SimpleTooltip from './simple-tooltip';
-import Commands, {Command} from './commands';
+import SimpleTooltip from '../atom/simple-tooltip';
+import Commands, {Command} from '../atom/commands';
 import FilePatchSelection from '../models/file-patch-selection';
 import Switchboard from '../switchboard';
 import RefHolder from '../models/ref-holder';
