+++ lib/views/git-timings-view.js
@@ -10,7 +10,7 @@ import {autobind} from 'core-decorators';
 import memoize from 'lodash.memoize';
 import fs from 'fs-extra';
 
-import Octicon from './octicon';
+import Octicon from '../atom/octicon';
 
 const genArray = memoize(function genArray(interval, count) {
   const arr = [];
