+++ lib/views/hunk-view.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {autobind} from 'core-decorators';
 
-import SimpleTooltip from './simple-tooltip';
+import SimpleTooltip from '../atom/simple-tooltip';
 import ContextMenuInterceptor from '../context-menu-interceptor';
 
 export default class HunkView extends React.Component {
