+++ lib/views/init-dialog.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import {autobind} from 'core-decorators';
 import {CompositeDisposable} from 'event-kit';
 
-import Commands, {Command} from './commands';
+import Commands, {Command} from '../atom/commands';
 
 export default class InitDialog extends React.Component {
   static propTypes = {
