+++ lib/views/issueish-badge.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
 
-import Octicon from './octicon';
+import Octicon from '../atom/octicon';
 
 const typeAndStateToIcon = {
   Issue: {
