+++ lib/views/issueish-timeline-view.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import {autobind} from 'core-decorators';
 
 import {RelayConnectionPropType} from '../prop-types';
-import Octicon from '../views/octicon';
+import Octicon from '../atom/octicon';
 import CommitsContainer from './../containers/timeline-items/commits-container.js';
 import IssueCommentContainer from './../containers/timeline-items/issue-comment-container.js';
 import MergedEventContainer from './../containers/timeline-items/merged-event-container.js';
