+++ lib/views/push-pull-view.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import cx from 'classnames';
 
 import {RemotePropType, BranchPropType} from '../prop-types';
-import Tooltip from './tooltip';
+import Tooltip from '../atom/tooltip';
 import RefHolder from '../models/ref-holder';
 
 function getIconClass(icon, animation) {
