+++ lib/views/staging-view.js
@@ -15,7 +15,7 @@ import MergeConflictListItemView from './merge-conflict-list-item-view';
 import CompositeListSelection from '../models/composite-list-selection';
 import ResolutionProgress from '../models/conflicts/resolution-progress';
 import FilePatchController from '../controllers/file-patch-controller';
-import Commands, {Command} from './commands';
+import Commands, {Command} from '../atom/commands';
 
 const debounce = (fn, wait) => {
   let timeout;
