+++ test/atom/decoration.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import sinon from 'sinon';
 import {mount} from 'enzyme';
 
-import Decoration from '../../lib/views/decoration';
+import Decoration from '../../lib/atom/decoration';
 
 describe('Decoration', () => {
   let atomEnv, editor, marker;
